refactor: chat folder structure
