+++ src/app/(admin)/new_dashboard/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./ChatHeader";
+export * from "./ContactList";
+export * from "./MessageInput";
+export * from "./MessageList";
