+++ src/app/(admin)/new_dashboard/types/index.ts
@@ -1,0 +1,11 @@
+export type Message = {
+  sender: string;
+  text: string;
+  timestamp: string;
+};
+
+export type Contact = {
+  name: string;
+  avatar: string;
+  lastMessage: string;
+};
